Changed all entities clases

+++ src/main/resources/schema.sql
@@ -40,34 +40,34 @@ END $$;
 
 -- Tabla PROVINCIAS
 CREATE TABLE IF NOT EXISTS provincias (
-                                          id BIGSERIAL PRIMARY KEY,
+                                          id SERIAL PRIMARY KEY,
                                           nombre VARCHAR(255) NOT NULL
     );
 
 -- Tabla USUARIOS
 CREATE TABLE IF NOT EXISTS usuarios (
-                                        id BIGSERIAL PRIMARY KEY,
+                                        id SERIAL PRIMARY KEY,
                                         nombre VARCHAR(255) NOT NULL,
     apellido VARCHAR(255) NOT NULL,
     tipo_documento tipo_documento NOT NULL,
     nro_doc VARCHAR(20) NOT NULL UNIQUE,
     rol rol_usuario NOT NULL,
-    provincia_id BIGINT NOT NULL,
+    provincia_id INTEGER NOT NULL,
     puntos INTEGER DEFAULT 0,
     FOREIGN KEY (provincia_id) REFERENCES provincias(id)
     );
 
 -- Tabla CUENTAS_BANCARIAS
 CREATE TABLE IF NOT EXISTS cuentas_bancarias (
-                                                 id BIGSERIAL PRIMARY KEY,
+                                                 id SERIAL PRIMARY KEY,
                                                  cbu BIGINT NOT NULL,
-                                                 id_usuario BIGINT NOT NULL,
+                                                 id_usuario INTEGER NOT NULL,
                                                  FOREIGN KEY (id_usuario) REFERENCES usuarios(id)
     );
 
 -- Tabla SPONSORS
 CREATE TABLE IF NOT EXISTS sponsors (
-                                        id BIGSERIAL PRIMARY KEY,
+                                        id SERIAL PRIMARY KEY,
                                         nombre VARCHAR(255) NOT NULL,
     ruta_img1 VARCHAR(500),
     ruta_img2 VARCHAR(500)
@@ -75,12 +75,12 @@ CREATE TABLE IF NOT EXISTS sponsors (
 
 -- Tabla EVENTOS
 CREATE TABLE IF NOT EXISTS eventos (
-                                       id BIGSERIAL PRIMARY KEY,
+                                       id SERIAL PRIMARY KEY,
                                        tipo tipo_evento NOT NULL,
                                        estado estado_evento NOT NULL,
-                                       organizador_id BIGINT NOT NULL,
-                                       cuenta_bancaria_id BIGINT,
-                                       provincia_id BIGINT NOT NULL,
+                                       organizador_id INTEGER NOT NULL,
+                                       cuenta_bancaria_id INTEGER,
+                                       provincia_id INTEGER NOT NULL,
                                        nombre VARCHAR(255) NOT NULL,
     descripcion TEXT,
     ruta_img VARCHAR(500),
@@ -90,7 +90,7 @@ CREATE TABLE IF NOT EXISTS eventos (
     direccion VARCHAR(500),
     costo_interno DECIMAL(10,2),
     costo_inscripcion DECIMAL(10,2),
-    sponsor_id BIGINT,
+    sponsor_id INTEGER,
     FOREIGN KEY (organizador_id) REFERENCES usuarios(id),
     FOREIGN KEY (cuenta_bancaria_id) REFERENCES cuentas_bancarias(id),
     FOREIGN KEY (provincia_id) REFERENCES provincias(id),
@@ -99,27 +99,27 @@ CREATE TABLE IF NOT EXISTS eventos (
 
 -- Tabla INSCRIPCIONES
 CREATE TABLE IF NOT EXISTS inscripciones (
-                                             id BIGSERIAL PRIMARY KEY,
-                                             usuario_id BIGINT NOT NULL,
+                                             id SERIAL PRIMARY KEY,
+                                             usuario_id INTEGER NOT NULL,
                                              estado estado_inscripcion NOT NULL,
-                                             evento_id BIGINT NOT NULL,
+                                             evento_id INTEGER NOT NULL,
                                              FOREIGN KEY (usuario_id) REFERENCES usuarios(id),
     FOREIGN KEY (evento_id) REFERENCES eventos(id)
     );
 
 -- Tabla PAGOS_USUARIOS
 CREATE TABLE IF NOT EXISTS pagos_usuarios (
-                                              id BIGSERIAL PRIMARY KEY,
-                                              id_usuario BIGINT NOT NULL,
-                                              id_evento BIGINT NOT NULL,
+                                              id SERIAL PRIMARY KEY,
+                                              id_usuario INTEGER NOT NULL,
+                                              id_evento INTEGER NOT NULL,
                                               FOREIGN KEY (id_usuario) REFERENCES usuarios(id),
     FOREIGN KEY (id_evento) REFERENCES eventos(id)
     );
 
 -- Tabla PETICIONES_ORGANIZADOR
 CREATE TABLE IF NOT EXISTS peticiones_organizadores (
-                                                        id BIGSERIAL PRIMARY KEY,
-                                                        usuario_id BIGINT NOT NULL,
+                                                        id SERIAL PRIMARY KEY,
+                                                        usuario_id INTEGER NOT NULL,
                                                         estado_peticion estado_peticion NOT NULL,
                                                         mensaje_usuario TEXT,
                                                         FOREIGN KEY (usuario_id) REFERENCES usuarios(id)
@@ -127,18 +127,21 @@ CREATE TABLE IF NOT EXISTS peticiones_organizadores (
 
 -- Tabla CANJEABLES
 CREATE TABLE IF NOT EXISTS canjeables (
-                                          id BIGSERIAL PRIMARY KEY,
-                                          id_sponsor BIGINT NOT NULL,
+                                          id SERIAL PRIMARY KEY,
+                                          id_sponsor INTEGER NOT NULL,
                                           link_pdf_drive VARCHAR(500),
     costo_puntos INTEGER NOT NULL,
     FOREIGN KEY (id_sponsor) REFERENCES sponsors(id)
     );
 
+-- Eliminar tabla sponsors_eventos ya que ahora es relación uno a uno
+-- La relación se maneja directamente en la tabla eventos con sponsor_id
+
 -- Tabla DONACIONES
 CREATE TABLE IF NOT EXISTS donaciones (
-                                          id BIGSERIAL PRIMARY KEY,
-                                          id_evento BIGINT NOT NULL,
-                                          id_usuario BIGINT NOT NULL,
+                                          id SERIAL PRIMARY KEY,
+                                          id_evento INTEGER NOT NULL,
+                                          id_usuario INTEGER NOT NULL,
                                           monto DECIMAL(10,2) NOT NULL,
     mensaje TEXT,
     FOREIGN KEY (id_evento) REFERENCES eventos(id),
